requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Django==4.2.16
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
fastapi==0.115.10
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
fonttools==4.55.5
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.3
matplotlib==3.10.0
motor==3.7.0
multidict==6.4.3
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==20.8
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
visitor==0.1.3
Werkzeug==2.3.7
yarl==1.20.0
